Change model parameter to fp16

+++ training/segmentation_training.py
@@ -700,7 +700,7 @@ def main():
 
     model_cls = MaskGitTransformer if config.model.get("architecture", "transformer") == "transformer" else MaskGiTUViT
     if config.model.get("pretrained_model_path", None) is not None:
-        model = model_cls.from_pretrained(config.model.pretrained_model_path)
+        model = model_cls.from_pretrained(config.model.pretrained_model_path, torch_dtype=torch.float16)
     else:
         model = model_cls(**config.model.transformer)
     mask_id = model.config.mask_token_id
